python_backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
langchain==0.1.0
langchain-google-genai==0.0.8
langchain-community==0.0.13
aiofiles==23.2.1
python-multipart==0.0.6
websockets==12.0
cors==1.0.1
